+++ Porters Park/MHSystems/app/src/main/java/com/mh/systems/porterspark/models/DeleteToken/DeleteTokenAPI.java
@@ -1,0 +1,117 @@
+
+package com.mh.systems.porterspark.models.DeleteToken;
+
+import com.google.gson.annotations.Expose;
+import com.google.gson.annotations.SerializedName;
+
+public class DeleteTokenAPI {
+
+    @SerializedName("aClientId")
+    @Expose
+    private String aClientId;
+    @SerializedName("aCommand")
+    @Expose
+    private String aCommand;
+    @SerializedName("aJsonParams")
+    @Expose
+    private AJsonParamsDeleteToken aJsonParams;
+    @SerializedName("aModuleId")
+    @Expose
+    private String aModuleId;
+    @SerializedName("aUserClass")
+    @Expose
+    private String aUserClass;
+
+    /**
+     * No args constructor for use in serialization
+     */
+    public DeleteTokenAPI() {
+    }
+
+    /**
+     * @param aJsonParams
+     * @param aModuleId
+     * @param aUserClass
+     * @param aClientId
+     * @param aCommand
+     */
+    public DeleteTokenAPI(String aClientId, String aCommand, AJsonParamsDeleteToken aJsonParams, String aModuleId, String aUserClass) {
+        super();
+        this.aClientId = aClientId;
+        this.aCommand = aCommand;
+        this.aJsonParams = aJsonParams;
+        this.aModuleId = aModuleId;
+        this.aUserClass = aUserClass;
+    }
+
+    /**
+     * @return The aClientId
+     */
+    public String getAClientId() {
+        return aClientId;
+    }
+
+    /**
+     * @param aClientId The aClientId
+     */
+    public void setAClientId(String aClientId) {
+        this.aClientId = aClientId;
+    }
+
+    /**
+     * @return The aCommand
+     */
+    public String getACommand() {
+        return aCommand;
+    }
+
+    /**
+     * @param aCommand The aCommand
+     */
+    public void setACommand(String aCommand) {
+        this.aCommand = aCommand;
+    }
+
+    /**
+     * @return The aJsonParams
+     */
+    public AJsonParamsDeleteToken getAJsonParams() {
+        return aJsonParams;
+    }
+
+    /**
+     * @param aJsonParams The aJsonParams
+     */
+    public void setAJsonParams(AJsonParamsDeleteToken aJsonParams) {
+        this.aJsonParams = aJsonParams;
+    }
+
+    /**
+     * @return The aModuleId
+     */
+    public String getAModuleId() {
+        return aModuleId;
+    }
+
+    /**
+     * @param aModuleId The aModuleId
+     */
+    public void setAModuleId(String aModuleId) {
+        this.aModuleId = aModuleId;
+    }
+
+    /**
+     * @return The aUserClass
+     */
+    public String getAUserClass() {
+        return aUserClass;
+    }
+
+    /**
+     * @param aUserClass The aUserClass
+     */
+    public void setAUserClass(String aUserClass) {
+        this.aUserClass = aUserClass;
+    }
+
+}