+++ Porters Park/MHSystems/app/src/main/java/com/mh/systems/porterspark/web/api/WebServiceMethods.java
@@ -4,6 +4,8 @@ import com.google.gson.JsonObject;
 import com.mh.systems.porterspark.models.AddMemberAPI;
 import com.mh.systems.porterspark.models.ClubNews.ClubNewsAPI;
 import com.mh.systems.porterspark.models.ClubNews.ClubNewsDetailAPI;
+import com.mh.systems.porterspark.models.ClubNewsThumbnail.ClubNewsThumbnailAPI;
+import com.mh.systems.porterspark.models.ClubNewsThumbnail.ClubNewsThumbnailDetailAPI;
 import com.mh.systems.porterspark.models.CompetitionResultAPI;
 import com.mh.systems.porterspark.models.CompetitionJoinAPI;
 import com.mh.systems.porterspark.models.CompetitionUnjoinAPI;
@@ -12,6 +14,7 @@ import com.mh.systems.porterspark.models.ContactUs.ContactUsAPI;
 import com.mh.systems.porterspark.models.CourseDiaryAPI;
 import com.mh.systems.porterspark.models.CourseNames.CourseNamesAPI;
 import com.mh.systems.porterspark.models.DashboardAPI;
+import com.mh.systems.porterspark.models.DeleteToken.DeleteTokenAPI;
 import com.mh.systems.porterspark.models.EditDetailMode.EditDetailModeAPI;
 import com.mh.systems.porterspark.models.ForgotPassword.ForgotPasswordAPI;
 import com.mh.systems.porterspark.models.Friends.RemoveFriendAPI;
@@ -23,10 +26,13 @@ import com.mh.systems.porterspark.models.MembersDetailAPI;
 import com.mh.systems.porterspark.models.FinanceAPI;
 import com.mh.systems.porterspark.models.ResetPassword.ResetPasswordAPI;
 import com.mh.systems.porterspark.models.TogglePrivacy.TogglePrivacyAPI;
+import com.mh.systems.porterspark.models.UnreadNewsCount.GetUnreadNewsCountAPI;
 import com.mh.systems.porterspark.models.UpdatePassword.UpdatePassswordAPI;
 import com.mh.systems.porterspark.models.competitionsEntry.CompEligiblePlayersAPI;
 import com.mh.systems.porterspark.models.competitionsEntry.GetClubEventAPI;
 import com.mh.systems.porterspark.models.competitionsEntry.UpdateCompEntryAPI;
+import com.mh.systems.porterspark.models.featuresflag.FeatureFlagsAPI;
+import com.mh.systems.porterspark.models.registerToken.RegisterTokenAPI;
 
 import retrofit.Callback;
 import retrofit.http.Body;
@@ -388,5 +394,76 @@ public interface WebServiceMethods {
      */
     @GET("/api/ApifsiGateway/TopUps")
     public void getTopUpPricesList(@Query("aClientId") String aClientId, @Query("aMemberId") String aMemberId, Callback<JsonObject> response);
+
+    /**
+     * Declaration of DELETE TOKEN API which is using for push
+     * notifications.
+     * <p/>
+     * TYPE : POST
+     *
+     * @param deleteTokenAPI : Delete Token which sending at time of Registeration.
+     * @param response       : Weather api response in JSON format.
+     */
+    @POST("/webapi/api/ClubsApp")
+    public void deleteToken(@Body DeleteTokenAPI deleteTokenAPI, Callback<JsonObject> response);
+
+    /**
+     * Declaration of GET UNREAD CLUB NEWS API which will be
+     * display on dashboard at top of Club News icon.
+     * <p/>
+     * TYPE : POST
+     *
+     * @param getUnreadNewsCountAPI : Object of unread club news.
+     * @param response              : Weather api response in JSON format.
+     */
+    @POST("/webapi/api/ClubsApp")
+    public void getUnreadClubNewsCount(@Body GetUnreadNewsCountAPI getUnreadNewsCountAPI, Callback<JsonObject> response);
+
+    /**
+     * Declaration of REGISTRATION TOKEN API in background service.
+     * <p/>
+     * TYPE : POST
+     *
+     * @param registerTokenAPI : Time hour in running device.
+     * @param response         : Weather api response in JSON format.
+     */
+    @POST("/webapi/api/ClubsApp")
+    public void registerToken(@Body RegisterTokenAPI registerTokenAPI, Callback<JsonObject> response);
+
+    /**
+     * Declaration of Club News web service method with Thumbnail
+     * of image.
+     * <p/>
+     * TYPE : POST
+     *
+     * @param clubNewsThumbnailAPI :  Club News with Thumbnail.
+     * @param response
+     */
+    @POST("/webapi/api/ClubsApp")
+    public void getClubNewsThumbnail(@Body ClubNewsThumbnailAPI clubNewsThumbnailAPI, Callback<JsonObject> response);
+
+    /**
+     * Declaration of Club News detail content of Thumbnail web
+     * service method.
+     * <p/>
+     * TYPE : POST
+     *
+     * @param clubNewsThumbnailDetailAPI :  Club News detail of Thumbnail.
+     * @param response
+     */
+    @POST("/webapi/api/ClubsApp")
+    public void getClubNewsThumbnailDetail(@Body ClubNewsThumbnailDetailAPI clubNewsThumbnailDetailAPI, Callback<JsonObject> response);
+
+    /**
+     * Call Features flag web service to get list of
+     * features show on dashboard.
+     * <p/>
+     * TYPE : POST
+     *
+     * @param featureFlagsAPI       : Pass instance of features flag.
+     * @param response              : Response in JSON format.
+     */
+    @POST("/webapi/api/ClubsApp")
+    public void getFeaturesFlagOptions(@Body FeatureFlagsAPI featureFlagsAPI, Callback<JsonObject> response);
 }