+++ Porters Park/MHSystems/CHANGELOG.md
@@ -10,6 +10,15 @@
     #####2nd Feb, 2017
      - UPDATE: Replace 'SimpleDrawerView' with 'CircleImageView' for better performance.
 
+     #####21st Feb, 2017
+    - Enhanced: Remove 'libValidation' library used for input passwords validations and implements custom code.
+    - UPDATE: Display 'No Image' till club news doesn't load image.
+    - UPDATE: Club News with thumbnail feature added.
+    - Enhanced: Move 'Features Flag' and 'Registration Token' from Login to Dashboard.
+    - NEW: Integrate Club News Push Notification [Hide].
+    - NEW: Add 'Delete Token' functionality.
+    - NEW: Add 'Unread News Badger' count functionality [Hide].
+
 ##Version 1.4.0 (2016-12-02)
 
 - New: Contact us feature added.